requirements.txt
pydantic
PyYAML
simplegmail
openai